fga-eps-mds/2018.2-IndicaAi#127 - Creating RegisterAPIForm component and removing the form from RegisterlocalAPI container
