+++ mobile/screens/tab_navigator/indica_ai/containers/RegisterLocal.js
@@ -177,7 +177,6 @@ const styles = StyleSheet.create({
     }
   },
   registerButton: {
-    color: '#0AACCC',
     top: 30,
     marginHorizontal: 10
   }
